remove old nodes
