+++ Dockerfile
@@ -82,12 +82,12 @@ echo "Sensor Mode: Uses config/sensors_sim.json by default"\n\
 echo ""\n\
 echo "Available commands:"\n\
 echo "  # Simulation mode (default):"\n\
-echo "  ros2 launch embr embr_launch_v2.py config_file:=config/sensors_sim.json"\n\
+echo "  ros2 launch embr embr_launch.py config_file:=config/sensors_sim.json"\n\
 echo ""\n\
 echo "  # Individual nodes:"\n\
-echo "  ros2 run embr getTemp_v2 --ros-args -p config_file:=config/sensors_sim.json"\n\
-echo "  ros2 run embr getCube_v2 --ros-args -p config_file:=config/sensors_sim.json"\n\
-echo "  ros2 run embr sendRf_v2 --ros-args -p config_file:=config/sensors_sim.json"\n\
+echo "  ros2 run embr getTemp --ros-args -p config_file:=config/sensors_sim.json"\n\
+echo "  ros2 run embr getCube --ros-args -p config_file:=config/sensors_sim.json"\n\
+echo "  ros2 run embr sendRf --ros-args -p config_file:=config/sensors_sim.json"\n\
 echo ""\n\
 echo "  # Monitoring:"\n\
 echo "  ros2 topic list                       - List topics"\n\
