+++ ros2_ws/src/embr/config/CONFIG.md
@@ -43,10 +43,10 @@ Simply run nodes without any parameters to use `config/sensors.json`:
 
 ```bash
 # Single node
-ros2 run embr getTemp_v2
+ros2 run embr getTemp
 
 # All nodes via launch file
-ros2 launch embr embr_launch_v2.py
+ros2 launch embr embr_launch.py
 ```
 
 ### Using Custom Configuration
@@ -55,13 +55,13 @@ Pass a custom config file to override the default:
 
 ```bash
 # Single node with simulation config
-ros2 run embr getTemp_v2 --ros-args -p config_file:=src/embr/config/sensors_sim.json
+ros2 run embr getTemp --ros-args -p config_file:=src/embr/config/sensors_sim.json
 
 # All nodes with simulation config via launch file
-ros2 launch embr embr_launch_v2.py config_file:=config/sensors_sim.json
+ros2 launch embr embr_launch.py config_file:=config/sensors_sim.json
 
 # Use an external config file
-ros2 launch embr embr_launch_v2.py config_file:=/path/to/my/custom_sensors.json
+ros2 launch embr embr_launch.py config_file:=/path/to/my/custom_sensors.json
 ```
 
 ## Troubleshooting
