--- ros2_ws/src/embr/embr/getCube_v2.py
@@ -1,99 +1,0 @@
-"""
-Cube Orange GPS publisher node with sensor abstraction.
-Supports both real and simulated sensors.
-"""
-
-import rclpy
-from rclpy.node import Node
-import time
-from msg_interface.msg import GPSAndIMU
-from embr.sensors import create_sensor, SensorConfig, SensorFactory
-
-
-class AttitudePublisher(Node):
-    def __init__(self):
-        super().__init__('attitude_publisher')
-        
-        # Declare parameter for config file path
-        self.declare_parameter('config_file', '')
-        config_file = self.get_parameter('config_file').value
-        
-        # Use default config path if not provided
-        if not config_file:
-            config_file = 'src/embr/config/sensors.json'
-        
-        # Try to load from config file
-        try:
-            configs = SensorFactory.load_config(config_file)
-            config = configs.get('cube')
-            
-            if config is None:
-                self.get_logger().error(f'Cube sensor not found in config file: {config_file}')
-                raise ValueError(f'Cube sensor configuration not found in {config_file}')
-            
-            self.get_logger().info(f'Loaded cube config from {config_file}')
-        except Exception as e:
-            self.get_logger().error(f'Failed to load config file: {e}')
-            raise
-        
-        # Create sensor
-        try:
-            self.sensor = create_sensor('cube', config)
-            self.sensor.start()
-
-            sensor_type = 'simulated' if 'Sim' in self.sensor.__class__.__name__ else 'real'
-            self.get_logger().info(f'Cube sensor initialized in {config.mode} mode (using {sensor_type} sensor)')
-        except Exception as e:
-            self.get_logger().error(f'Failed to initialize sensor: {e}')
-            raise
-
-        # Create publisher and timer (GPS + IMU message)
-        self.publisher_ = self.create_publisher(GPSAndIMU, 'gps', 10)
-        self.timer = self.create_timer(1.0, self.publish_attitude)
-
-    def publish_attitude(self):
-        """Read GPS data and publish."""
-        try:
-            gps_data = self.sensor.read()
-
-            gps_msg = GPSAndIMU()
-            # Core GPS fields
-            gps_msg.lat = float(gps_data.lat)
-            gps_msg.lon = float(gps_data.lon)
-            gps_msg.alt = float(gps_data.alt)
-            gps_msg.vel = float(gps_data.vel)
-            # IMU fields
-            gps_msg.pitch = float(gps_data.pitch)
-            gps_msg.yaw = float(gps_data.yaw)
-            gps_msg.roll = float(gps_data.roll)
-
-            self.get_logger().info(
-                f'GPS: Lat: {gps_msg.lat} Lon: {gps_msg.lon} '
-                f'Alt: {gps_msg.alt} Vel: {gps_msg.vel:.2f}'
-            )
-
-            self.publisher_.publish(gps_msg)
-        except Exception as e:
-            self.get_logger().error(f'Error reading GPS: {e}')
-
-    def destroy_node(self):
-        """Cleanup sensor on shutdown."""
-        if hasattr(self, 'sensor'):
-            self.sensor.stop()
-        super().destroy_node()
-
-
-def main(args=None):
-    rclpy.init(args=args)
-    attitude_publisher = AttitudePublisher()
-    try:
-        rclpy.spin(attitude_publisher)
-    except KeyboardInterrupt:
-        pass
-    finally:
-        attitude_publisher.destroy_node()
-        rclpy.shutdown()
-
-
-if __name__ == '__main__':
-    main()
