+++ ros2_ws/src/embr/setup.py
@@ -27,9 +27,6 @@ setup(
             'getTemp = embr.getTemp:main',
             'sendRf = embr.sendRf:main',
             'thermalStream = embr.thermalStream:main',
-            'getCube_v2 = embr.getCube_v2:main',
-            'getTemp_v2 = embr.getTemp_v2:main',
-            'sendRf_v2 = embr.sendRf_v2:main',
             'hotspotLocator = embr.hotspotLocator:main',
         ],
     },
